Move client mount retrieval to new system
